IMPLEMENTED MAIN PROGRAM WITH USER INPUT AND OUTPUT
Added the main program logic to prompt users for a year input between 1980 and 2024. The program retrieves and displays the corresponding Stanley Cup champion from the pre-configured HashMap of champions. It handles invalid input with error messages and provides an option to continue or exit based on user input. Program runs as it should. Added a README to explain the program's purpose and how to use it.

SIGNED-OFF-BY: Jani Hiltunen
<152394729+JangoHiltzy@users.noreply.github.com>